refactor(agents): own controller startup runtime
