+++ docs/jangar/architecture-inventory.md
@@ -22,19 +22,19 @@ This file is generated by `bun run --cwd services/jangar docs:inventory:write`.
 | 5    | `services/jangar/src/server/orchestration-controller.ts`           | 2138 |
 | 6    | `services/jangar/src/server/chat-completion-encoder.ts`            | 2076 |
 | 7    | `services/jangar/src/server/torghut-market-context-agents.ts`      | 1981 |
-| 8    | `services/jangar/src/server/agents-controller/index.ts`            | 1824 |
-| 9    | `services/jangar/src/server/torghut-simulation-control-plane.ts`   | 1741 |
-| 10   | `services/jangar/src/server/github-review-store.ts`                | 1459 |
-| 11   | `services/jangar/src/server/codex-judge-store.ts`                  | 1329 |
-| 12   | `services/jangar/src/server/torghut-whitepapers.ts`                | 1328 |
-| 13   | `services/jangar/src/server/chat.ts`                               | 1313 |
-| 14   | `services/jangar/src/server/agentctl-grpc.ts`                      | 1171 |
-| 15   | `services/jangar/src/routes/github/pulls/$owner/$repo/$number.tsx` | 1028 |
-| 16   | `services/jangar/src/server/torghut-trading.ts`                    |  988 |
-| 17   | `services/jangar/src/server/db.ts`                                 |  974 |
-| 18   | `services/jangar/src/routes/torghut/trading.tsx`                   |  952 |
-| 19   | `services/jangar/src/server/torghut-quant-metrics.ts`              |  929 |
-| 20   | `services/jangar/src/server/github-review-actions.ts`              |  928 |
+| 8    | `services/jangar/src/server/torghut-simulation-control-plane.ts`   | 1741 |
+| 9    | `services/jangar/src/server/github-review-store.ts`                | 1459 |
+| 10   | `services/jangar/src/server/codex-judge-store.ts`                  | 1329 |
+| 11   | `services/jangar/src/server/torghut-whitepapers.ts`                | 1328 |
+| 12   | `services/jangar/src/server/chat.ts`                               | 1313 |
+| 13   | `services/jangar/src/server/agentctl-grpc.ts`                      | 1171 |
+| 14   | `services/jangar/src/routes/github/pulls/$owner/$repo/$number.tsx` | 1028 |
+| 15   | `services/jangar/src/server/torghut-trading.ts`                    |  988 |
+| 16   | `services/jangar/src/server/db.ts`                                 |  974 |
+| 17   | `services/jangar/src/routes/torghut/trading.tsx`                   |  952 |
+| 18   | `services/jangar/src/server/torghut-quant-metrics.ts`              |  929 |
+| 19   | `services/jangar/src/server/github-review-actions.ts`              |  928 |
+| 20   | `services/jangar/src/server/github-client.ts`                      |  905 |
 
 ## Control-Plane Route Surface
