+++ services/agents/package.json
@@ -6,20 +6,24 @@
   "exports": {
     "./server/controller-startup-retry": "./src/server/controller-startup-retry.ts",
     "./server/env-compat": "./src/server/env-compat.ts",
+    "./server/feature-flags": "./src/server/feature-flags.ts",
     "./server/health": "./src/server/health.ts",
     "./server/http": "./src/server/http.ts",
     "./server/http-runtime": "./src/server/http-runtime.ts",
+    "./server/integrations-config": "./src/server/integrations-config.ts",
     "./server/kube-gateway": "./src/server/kube-gateway.ts",
     "./server/kube-types": "./src/server/kube-types.ts",
     "./server/kube-watch": "./src/server/kube-watch.ts",
     "./server/kubernetes-bun-transport": "./src/server/kubernetes-bun-transport.ts",
     "./server/kubernetes-watch-client": "./src/server/kubernetes-watch-client.ts",
+    "./server/metrics": "./src/server/metrics.ts",
     "./server/namespace-scope": "./src/server/namespace-scope.ts",
     "./server/primitives": "./src/server/primitives.ts",
     "./server/primitives-policy": "./src/server/primitives-policy.ts",
     "./server/ready": "./src/server/ready.ts",
     "./server/runtime-identity": "./src/server/runtime-identity.ts",
     "./server/status-utils": "./src/server/status-utils.ts",
+    "./server/agents-controller": "./src/server/agents-controller/index.ts",
     "./server/agents-controller/agent-run-reconciler": "./src/server/agents-controller/agent-run-reconciler.ts",
     "./server/agents-controller/mutable-state": "./src/server/agents-controller/mutable-state.ts",
     "./server/agents-controller/rate-limits": "./src/server/agents-controller/rate-limits.ts",
